refactor : 컴포넌트 별로 폴더 및 파일 분리
