+++ client/components/common/Button.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 박혜정
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-12-06
- */
-
 import { MouseEventHandler } from 'react';
 
 type ButtonText = {
