+++ client/components/common/Header/Logo.tsx
@@ -1,10 +1,3 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-12-03
- */
-
-import Image from 'next/image';
 import Link from 'next/link';
 
 export const Logo = () => {
