+++ client/components/common/Header/SearchBar.tsx
@@ -1,13 +1,7 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-12-04
- */
-
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons';
 import { useSetRecoilState } from 'recoil';
-import { keywordAtom } from '../../atomsYW';
+import { keywordAtom } from '../../../atomsYW';
 import { useRouter } from 'next/router';
 import { SubmitHandler, useForm } from 'react-hook-form';
