+++ client/components/common/MorakLogo.tsx
