+++ client/components/common/Pagination.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 박혜정
- * 최초 작성일: 2022-12-02
- * 최근 수정일: 2022-12-02
- */
-
 import { useRecoilState } from 'recoil';
 import { currPagesIndexAtom } from '../../atomsHJ';
