+++ client/components/common/QuillEditor/index.tsx
@@ -1,11 +1,5 @@
-/*
- * 책임 작성자: 박혜정
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-11-27
- */
-
 import dynamic from 'next/dynamic';
-import { Loader } from '../common/Loader';
+import { Loader } from '../Loader';
 
 const formats = [
   'header',
