+++ client/components/common/SocialLoginBtn.tsx
@@ -1,10 +1,3 @@
-/*
- * 책임 작성자: 정하승
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-12-01(박혜정)
- * 개요: 소셜 로그인 버튼에 대한 컴포넌트입니다.
- */
-
 import { faComment } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
