+++ client/components/main/Calendar.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-18
- * 최근 수정일: 2022-12-07
- */
-
 import FullCalendar from '@fullcalendar/react';
 import daygridPlugin from '@fullcalendar/daygrid';
