+++ client/components/main/CarouselMain.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-18
- * 최근 수정일: 2022-11-18
- */
-
 import * as React from 'react';
 import { useState, useEffect } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
