+++ client/components/main/HallOfFame.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-18
- * 최근 수정일: 2022-12-02
- */
-
 import Image from 'next/image';
 import Link from 'next/link';
 import { useEffect, useState } from 'react';
