+++ client/components/questions/SearchWithTagButton/index.tsx
@@ -1,9 +1,3 @@
-/*
- * 책임 작성자: 정하승
- * 최초 작성일: 2022-11-19
- * 최근 수정일: 2022-12-03(박혜정)
- */
-
 const tags = [
   { tagId: 0, name: 'JAVA' },
   { tagId: 1, name: 'C' },
