+++ client/pages/index.tsx
@@ -1,22 +1,17 @@
-/*
- * 책임 작성자: 박연우
- * 최초 작성일: 2022-11-14
- * 최근 수정일: 2022-11-18
- */
-
 import { NextPage } from 'next';
 import { Footer } from '../components/common/Footer';
 import { Header } from '../components/common/Header';
-import { CarouselMain } from '../components/yeonwoo/CarouselMain';
-import { CheerUp } from '../components/yeonwoo/CheerUp';
-import { HallOfFame } from '../components/yeonwoo/HallOfFame';
-import { ListLately } from '../components/yeonwoo/ListLately';
+
 import dynamic from 'next/dynamic';
 import { Seo } from '../components/common/Seo';
 import Link from 'next/link';
+import { HallOfFame } from '../components/main/HallOfFame';
+import { ListLately } from '../components/main/ListLately';
+import { CheerUp } from '../components/main/CheerUp';
+import { CarouselMain } from '../components/main/CarouselMain';
 
 const Calendar = dynamic<any>(
-  () => import('../components/yeonwoo/Calendar').then((mod) => mod.Calendar),
+  () => import('../components/main/Calendar').then((mod) => mod.Calendar),
   {
     ssr: false,
   },
